# openapi.yaml
openapi: "3.1.0"
info:
  title: Orchestrator Platform API
  description: |
    Multi-Model AI Agent Orchestration Platform API.
    Provides a unified gateway to multiple LLM providers (OpenAI, Anthropic, Google),
    isolated code execution sandboxes, and a workflow orchestration engine that
    decomposes objectives into DAGs of parallel tasks.
  version: "0.1.0"
  contact:
    name: Orchestrator Platform
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key authentication. Pass your key as `Bearer sk-...`
  schemas:
...
    WorkflowConfig:
      type: object
      required: [objective]
      properties:
        objective:
          type: string
          maxLength: 10000
        model_overrides:
          type: object
          additionalProperties:
            type: string
        tools:
          type: array
          items:
            type: string
        max_credits:
          type: number
        callback_url:
          type: string
          format: uri
        human_approval:
          type: boolean
          default: false
        background:
          type: boolean
          default: false
        context_files:
          type: array
          items:
            type: object
            properties:
              filename:
                type: string
              content_base64:
                type: string
              media_type:
                type: string
    WorkflowSummary:
      type: object
      properties:
        id:
          type: string
        objective:
          type: string
        status:
          type: string
          enum: [pending, planning, executing, paused, completed, failed, cancelled]
        plan:
          type: object
          nullable: true
        credits_consumed:
          type: number
        created_at:
          type: string
        updated_at:
          type: string
        completed_at:
          type: string
          nullable: true
        tasks:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              task_type:
                type: string
              description:
                type: string
              status:
                type: string
              output_preview:
                type: string
                nullable: true
              started_at:
                type: string
                nullable: true
              completed_at:
                type: string
                nullable: true
paths:
...
  /v1/workflows:
    post:
      summary: Create and start a workflow
      description: |
        Takes a natural-language objective, uses an LLM to plan a DAG of tasks,
        and begins executing them in parallel.
      tags: [Workflows]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowConfig'
      responses:
        "201":
          description: Workflow created and planning started
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflow_id:
                    type: string
                  status:
                    type: string
                  created_at:
                    type: string
                  plan:
                    type: object
                    properties:
                      task_count:
                        type: integer
                      tasks:
                        type: array
                        items:
                          type: object
    get:
      summary: List workflows
      tags: [Workflows]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Workflow list
          content:
            application/json:
              schema:
                type: object
                properties:
                  workflows:
                    type: array
                    items:
                      $ref: '#/components/schemas/WorkflowSummary'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
  /v1/workflows/{id}:
    get:
      summary: Get workflow details
      tags: [Workflows]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowSummary'
    delete:
      summary: Cancel a workflow
      tags: [Workflows]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workflow cancelled
  /v1/workflows/{id}/stream:
    get:
      summary: Stream workflow events (SSE)
      description: |
        Server-Sent Events stream of workflow progress.
        Events: planning_complete, task_started, task_completed, task_failed,
        human_approval_required, workflow_completed, workflow_failed, credit_update.
      tags: [Workflows]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: SSE event stream
          content:
            text/event-stream:
              schema:
                type: string
  /v1/workflows/{id}/approve:
    post:
      summary: Approve or reject a pending task
      tags: [Workflows]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [task_id, approved]
              properties:
                task_id:
                  type: string
                approved:
                  type: boolean
                feedback:
                  type: string
      responses:
        "200":
          description: Approval processed
...